requirements.txt
# Search Similar MCP Server Dependencies
# Generated from pyproject.toml for Docker deployment

httpx>=0.28.1
mcp[cli]>=1.14.0
numpy>=2.3.3
pgvector>=0.4.1
psycopg2>=2.9.10
pydantic>=2.11.9
python-dotenv>=1.1.1
requests>=2.32.5
sqlalchemy>=2.0.43
truststore